refactor: move some models outside of ServerData file
