+++ UnityProject/Assets/Scripts/Core/Database/Models/FireStore.cs
@@ -1,0 +1,32 @@
+﻿using System;
+
+namespace Core.Database.Models
+{
+	[Serializable]
+	public class FireStoreResponse
+	{
+		public FireStoreError error;
+		public string name;
+		public FireStoreFields fields;
+	}
+
+	[Serializable]
+	public class FireStoreError
+	{
+		public ushort code;
+		public string message;
+		public string status;
+	}
+
+	[Serializable]
+	public class FireStoreFields
+	{
+		public FireStoreCharacter character;
+	}
+
+	[Serializable]
+	public class FireStoreCharacter
+	{
+		public string stringValue;
+	}
+}
