+++ UnityProject/Assets/Scripts/Core/Database/Models/FireStore.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 15fd40635d1249a0bf25951ecfb1b229
+timeCreated: 1642066191
