+++ UnityProject/Assets/Scripts/Core/Database/Models/ServerStatus.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b3cd37f79ead4d6eafe8fbf564973fb1
+timeCreated: 1642064641
