+++ UnityProject/Assets/Scripts/Core/Database/ServerData.cs
@@ -163,32 +163,4 @@ namespace DatabaseAPI
 		public string refreshToken;
 		public string userID;
 	}
-
-	[Serializable]
-	public class FireStoreResponse
-	{
-		public FireStoreError error;
-		public string name;
-		public FireStoreFields fields;
-	}
-
-	[Serializable]
-	public class FireStoreError
-	{
-		public ushort code;
-		public string message;
-		public string status;
-	}
-
-	[Serializable]
-	public class FireStoreFields
-	{
-		public FireStoreCharacter character;
-	}
-
-	[Serializable]
-	public class FireStoreCharacter
-	{
-		public string stringValue;
-	}
 }
